UI v0.1
Update:
- Remove My Devices page
- Embed My Devices list to profile page
