+++ tlc-website copy/src/components/NavBar/Navbar.js
@@ -15,8 +15,8 @@ export default function Navbar() {
             </Link>
             <div className="nav-links">
                 {/* <CustomLink to="/album"> <i className="fas fa-image"></i> <span>Albums</span></CustomLink>
-                <CustomLink to="/film"> <i className="fas fa-film"></i> <span>Films</span></CustomLink> */}
-                <CustomLink to="/myDevices"> <i className="fas fa-video"> </i><span>Devices</span></CustomLink>
+                <CustomLink to="/film"> <i className="fas fa-film"></i> <span>Films</span></CustomLink> 
+                <CustomLink to="/myDevices"> <i className="fas fa-video"> </i><span>Devices</span></CustomLink>*/}
                 <CustomLink to="/PairDevices"> <i className="fas fa-link"> </i><span>Pair</span></CustomLink>
                 <CustomLink to="/about">< i className="fas fa-info-circle"> </i><span>About</span></CustomLink>
                 {userProfile.imageUrl && (
